+++ scripts/pypi.sh
@@ -1,0 +1,16 @@
+#! /bin/bash 
+###########################################
+# publish package to pypi
+###########################################
+
+# constants
+baseDir=$(cd `dirname "$0"`;pwd)
+export PYTHONUNBUFFERED=1
+export PATH=/opt/miniconda3/envs/venv-py3/bin:$PATH
+
+# functions
+
+# main 
+[ -z "${BASH_SOURCE[0]}" -o "${BASH_SOURCE[0]}" = "$0" ] || return
+cd $baseDir/..
+python setup.py upload -r pypi